+++ lib/src/image_file_reader.dart
@@ -3,5 +3,5 @@
 /// Use of this source code is governed by a BSD-style license that
 /// can be found in the LICENSE file.
 
-export 'config_file_reader_unsupported.dart'
-    if (dart.library.io) 'config_file_reader_io.dart';
+export 'image_file_reader_unsupported.dart'
+    if (dart.library.io) 'image_file_reader_io.dart';