+++ lib/src/image_file_reader_unsupported.dart
@@ -3,5 +3,5 @@
 /// Use of this source code is governed by a BSD-style license that
 /// can be found in the LICENSE file.
 
-export 'config_file_reader_unsupported.dart'
-    if (dart.library.io) 'config_file_reader_io.dart';
+Future<Map> loadImage(Uri imagePath) async =>
+    throw UnsupportedError('Loading images from disk is not supported on web.');