+++ lib/src/listing/mixins/gilded.dart
@@ -16,8 +16,14 @@ mixin GildedListingMixin {
   String get path;
 
   /// Returns a [Stream] of content that has been gilded.
-  Stream<UserContentInitialized> gilded({Map<String, String> params}) =>
+  ///
+  /// `limit` is the maximum number of objects returned by Reddit per request
+  /// (the default is 100). If provided, `after` specifies from which point
+  /// Reddit will return objects of the requested type. `params` is a set of
+  /// additional parameters that will be forwarded along with the request.
+  Stream<UserContentInitialized> gilded(
+          {int limit, String after, Map<String, String> params}) =>
       ListingGenerator.createBasicGenerator<UserContentInitialized>(
           reddit, path + 'gilded',
-          params: params);
+          limit: limit, after: after, params: params);
 }